Updated copyrights to 2011. Bumped version to 1.2.2b1. Fixed caps lock preventing Warp from activating. Don't Warp when in a fullscreen application such as a game.

+++ WarpEdgeView.m
@@ -3,7 +3,7 @@
 //  Warp
 //
 //  Created by Kent Sutherland on 2/13/08.
-//  Copyright 2008-2009 Kent Sutherland. All rights reserved.
+//  Copyright 2008-2011 Kent Sutherland. All rights reserved.
 //
 
 #import "WarpEdgeView.h"
@@ -211,11 +211,11 @@ extern OSStatus CGContextCopyWindowCaptureContentsToRect(CGContextRef ctx, CGRec
 	NSInteger count = _windowCount;
 	
 	if (count > 0) {
-		NSInteger outCount;
+		int outCount;
 		NSInteger cid = [NSApp contextID];
 		CGRect cgrect;
 		
-		NSInteger *list = malloc(sizeof(NSInteger) * count);
+		int *list = malloc(sizeof(int) * count);
 		CGSGetWorkspaceWindowList(_CGSDefaultConnection(), _workspace + 1, count, list, &outCount);
 		
 		NSSize screenSize = [[NSScreen mainScreen] frame].size;
